Moving things around. trying to find what i want
