+++ src/data.rs
@@ -1,10 +1,4 @@
-// pub mod crop;
-pub mod plant;
-// pub mod task;
-
 use diesel::prelude::*;
-
-
 
 pub struct DataMgr {
     pub conn: PgConnection,
@@ -13,7 +7,6 @@ pub struct DataMgr {
 impl DataMgr {
     pub fn new(db: String) -> Self {
         DataMgr {
-            
             conn: PgConnection::establish(&db).expect(&format!("Error connecting to {}", db)),
         }
     }
